pyproject.toml
[project]
name = "unwounded-ai"
version = "0.1.0"
description = "Unwounded AI runtime"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "litellm>=1.50.0",
  "langgraph>=0.2.40",
  "langgraph-checkpoint>=2.0.0",
  "langgraph-checkpoint-postgres>=2.0.0",
  "langsmith>=0.1.125",
  "braintrust>=0.0.165",
  "mem0ai>=0.1.0",
  "duckduckgo-search>=6.3.0",
  "tavily-python>=0.5.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "anyio>=4.4.0",
  "trio>=0.25.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
